+++ apps/transport/test/transport_web/live_views/gtfs_diff_select_live_test.exs
@@ -2,7 +2,7 @@ defmodule TransportWeb.Live.GTFSDiffSelectLiveTest do
   use ExUnit.Case, async: true
   import Phoenix.LiveViewTest
   doctest TransportWeb.Live.GTFSDiffSelectLive, import: true
-  doctest TransportWeb.Live.GTFSDiffSelectLive.Results, import: true
+  doctest TransportWeb.Live.GTFSDiffSelectLive.Differences, import: true
   doctest TransportWeb.Live.GTFSDiffSelectLive.Setup, import: true
 
   alias TransportWeb.GTFSDiffExplain
